# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "API Documentation",
        "license": {
            "name": "API"
        },
        "version": "1.0.0"
    },
    "basePath": "/public/api",
    "schemes": [
        "http",
        "https"
    ],
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "paths": {
...
        "/response/index": {
            "post": {
                "tags": [
                    "Response"
                ],
                "summary": "Get responses",
                "description": "Get responses",
                "operationId": "Get responses",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "type": "integer"
                    },
                    {
                        "name": "item_type",
                        "in": "query",
                        "description": "sort, chemical, seller, decorator",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "item_id",
                        "in": "query",
                        "description": "item id",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
        "/response/answers": {
            "post": {
                "tags": [
                    "Response"
                ],
                "summary": "Get  answers for  response  \u043f\u043e\u043b\u0443\u0447\u0438\u0442\u044c \u043a\u043e\u043c\u043c\u0435\u043d\u0442\u0430\u0440\u0438\u0438 \u043a \u043e\u0442\u0437\u044b\u0432\u0443",
                "description": "Get  answers for  response",
                "operationId": "Get  answers for  response",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "response id",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
        "/response/addresponse": {
            "post": {
                "tags": [
                    "Response"
                ],
                "summary": "Add  response",
                "description": "Add  response",
                "operationId": "Add  response",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "item_type",
                        "in": "query",
                        "description": "sort, chemical, seller, decorator",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "item_id",
                        "in": "query",
                        "description": "item id",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "rating",
                        "in": "query",
                        "description": "rating",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "text",
                        "in": "query",
                        "description": "response text",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "access_token": []
                    }
                ]
            }
        },
...
    },
...
    "securityDefinitions": {
        "access_token": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}